refactor(release): share component change-detection between release scripts
create-releases.mjs and release-pr-body.mjs both answered 'which components
changed vs base, and what are their notes?' with verbatim-duplicated
readPackage/readPackageFromGit helpers and parallel detection loops.

Extract the shared logic (plus extractReleaseNotes/escapeRegExp, which it
depends on) into releases.mjs and have both scripts consume
collectChangedComponents(). Guarantees the release-PR body advertises exactly
the set create-releases publishes, with one source of truth. Tests for the
moved functions follow them into releases.test.mjs.
